Added graphics
Added some smoothness to buttons and weather div borders, setted a new background for all pages and made all buttons the same with the cloudy background.

Added back buttons on MyEvents and NewEvent pages to be able to go back to the index page when the application is installed, therefore running fullscreen.

Added a background color change for the buttons in the MyEvents page based on the weather forecast for that day, and changed the way the date is displayed.

Fixed some minor errors like the race condition in the MyEvents page where it could get stuck in a while cycle.